Run the php-base tests. Add tests for version checking google/cloud libraries
