+++ builder/gen-dockerfile/tests/GenFilesCommandTest.php
@@ -188,97 +188,8 @@ class GenFilesCommandTest extends TestCase
                 '\\Google\\Cloud\\Runtimes\\Builder\\Exception\\InvalidComposerFlagsException'
             ],
             [
-                // stackdriver simple case
-                __DIR__ . '/test_data/stackdriver_simple',
-                null,
-                '',
-                '/app/web',
-                'added by the php runtime builder',
-                'gcr.io/google-appengine/php72:latest',
-                ["COMPOSER_FLAGS='--no-dev --prefer-dist' \\\n",
-                 "FRONT_CONTROLLER_FILE='index.php' \\\n",
-                 "DETECTED_PHP_VERSION='7.2' \\\n",
-                 "IS_BATCH_DAEMON_RUNNING='true' \n",
-                 "enable_stackdriver_integration.sh"
-                ]
-            ],
-            [
-                // stackdriver wildcard dep
-                __DIR__ . '/test_data/stackdriver_wildcard',
-                null,
-                '',
-                '/app/web',
-                'added by the php runtime builder',
-                'gcr.io/google-appengine/php72:latest',
-                ["COMPOSER_FLAGS='--no-dev --prefer-dist' \\\n",
-                 "FRONT_CONTROLLER_FILE='index.php' \\\n",
-                 "DETECTED_PHP_VERSION='7.2' \\\n",
-                 "IS_BATCH_DAEMON_RUNNING='true' \n",
-                 "enable_stackdriver_integration.sh"
-                ]
-            ],
-            [
-                // stackdriver individual packages
-                __DIR__ . '/test_data/stackdriver_individual',
-                null,
-                '',
-                '/app/web',
-                'added by the php runtime builder',
-                'gcr.io/google-appengine/php72:latest',
-                ["COMPOSER_FLAGS='--no-dev --prefer-dist' \\\n",
-                 "FRONT_CONTROLLER_FILE='index.php' \\\n",
-                 "DETECTED_PHP_VERSION='7.2' \\\n",
-                 "IS_BATCH_DAEMON_RUNNING='true' \n",
-                 "enable_stackdriver_integration.sh --individual"
-                ]
-            ],
-            [
-                // stackdriver old logging
-                __DIR__ . '/test_data/stackdriver_old_logging',
-                null,
-                '',
-                '/app/web',
-                'added by the php runtime builder',
-                'gcr.io/google-appengine/php72:latest',
-                [],
-                '\\Google\\Cloud\\Runtimes\\Builder\\Exception\\GoogleCloudVersionException'
-            ],
-            [
-                // stackdriver old error_reporting
-                __DIR__ . '/test_data/stackdriver_old_er',
-                null,
-                '',
-                '/app/web',
-                'added by the php runtime builder',
-                'gcr.io/google-appengine/php72:latest',
-                [],
-                '\\Google\\Cloud\\Runtimes\\Builder\\Exception\\GoogleCloudVersionException'
-            ],
-            [
                 // stackdriver no composer.json
                 __DIR__ . '/test_data/stackdriver_no_composer',
-                null,
-                '',
-                '/app/web',
-                'added by the php runtime builder',
-                'gcr.io/google-appengine/php72:latest',
-                [],
-                '\\Google\\Cloud\\Runtimes\\Builder\\Exception\\GoogleCloudVersionException'
-            ],
-            [
-                // stackdriver no google/cloud
-                __DIR__ . '/test_data/stackdriver_no_google_cloud',
-                null,
-                '',
-                '/app/web',
-                'added by the php runtime builder',
-                'gcr.io/google-appengine/php72:latest',
-                [],
-                '\\Google\\Cloud\\Runtimes\\Builder\\Exception\\GoogleCloudVersionException'
-            ],
-            [
-                // stackdriver old google/cloud
-                __DIR__ . '/test_data/stackdriver_old_google_cloud',
                 null,
                 '',
                 '/app/web',
