+++ php-base/stackdriver-files/enable_stackdriver_prepend.php
@@ -34,5 +34,6 @@ try {
         fclose($fp);
     }
 } catch (Exception $e) {
-    die($e->getMessage());
+    echo $e->getMessage();
+    exit(1);
 }
