+++ php-base/tests/bootstrap.php
@@ -1,1 +1,3 @@
 <?php
+
+require_once __DIR__ . '/../vendor/autoload.php';
