+++ php-base/tests/samples/mixed.json
@@ -1,7 +1,7 @@
 {
     "require": {
         "php": "7.0.*",
-        "ext-igbinary": "*",
-        "ext-mbstring": "*"
+        "ext-phalcon": "*",
+        "ext-sqlite3": "*"
     }
 }
