+++ php-base/tests/samples/phalcon.json
@@ -1,6 +1,6 @@
 {
     "require": {
         "php": "7.0.*",
-        "ext-igbinary": "*"
+        "ext-phalcon": "*"
     }
 }
