+++ php-base/tests/samples/shared.json
@@ -1,6 +1,6 @@
 {
     "require": {
         "php": "7.0.*",
-        "ext-mbstring": "*"
+        "ext-sqlite3": "*"
     }
 }
