+++ php-base/tests/samples/stackdriver_dev/composer.json
@@ -1,0 +1,11 @@
+{
+    "require": {
+        "google/cloud": "dev-master"
+    },
+    "repositories": [
+        {
+            "type": "git",
+            "url": "https://github.com/GoogleCloudPlatform/google-cloud-php"
+        }
+    ]
+}
