+++ php-base/tests/samples/stackdriver_old_er/composer.json
@@ -1,6 +1,6 @@
 {
     "require": {
-        "google/cloud-logging": "^1.4.1",
+        "google/cloud-logging": "*",
         "google/cloud-error-reporting": "^0.3.0"
     }
 }
